Reorganize components in subdirectories
