+++ src/App.jsx
@@ -4,8 +4,8 @@ import Profile from "./pages/Profile";
 import About from "./pages/About";
 import GenerateCalendar from "./pages/GenerateCalendar";
 import EditPreferences from "./pages/EditPreferences";
-import Navbar from "./components/Navbar";
-import Footer from "./components/Footer";
+import Navbar from "./components/common/Navbar.jsx";
+import Footer from "./components/common/Footer.jsx";
 import React from "react";
 import { UserProvider } from "./context/UserContext";
 import AdminDashboard from "./pages/AdminDashboard";
