+++ src/components/generate/FullDayEventItem.jsx
