+++ src/components/generate/RegularEventItem.jsx
