+++ src/pages/AdminCoursesDashboard.jsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useContext, useState} from "react";
 import api from "../api/axiosBackendConfig";
 import {UserContext} from "../context/UserContext";
-import CourseItem from "../components/CourseItem";
+import CourseItem from "../components/admin/CourseItem.jsx";
 import {Container, Row, Col, ListGroup, Button} from "react-bootstrap";
 import {useNavigate} from "react-router-dom";
 import {ClipLoader} from "react-spinners";
@@ -10,7 +10,7 @@ import {
     NO_PROBLEM,
 } from "../utill/Constants";
 import {toast} from "react-toastify";
-import PageHeader from "../components/PageHeader.jsx";
+import PageHeader from "../components/common/PageHeader.jsx";
 
 function AdminCoursesDashboard() {
     const [isUserAdmin, setIsUserAdmin] = useState(false);
@@ -132,7 +132,6 @@ function AdminCoursesDashboard() {
 
             <Row>
                 <Col>
-
                 </Col>
                 <Col className="text-end add-course-btn-col">
                     <Button variant="success" onClick={handleAddCourse}>
